Merge-Ready Commit *Hospital Address and Name info can be successfully retrieved from and saved to the patientInfo json file *Saving new hospital information to the json is done by longclicking an item in the nearby hospital listview and clicking yes on the dialog. *You can also click any item in this list to focus the map view on its location. *The Preferred hospital information is located above the nearby hospital list and can also be clicked. Once clicked, it also moves the map camera to it's position.
Other Changes:
*Made some ui changes in View Plans and Patient Edit. For some reason
the patient info button was obscured in the view plans screen, so I
moved it up a bit and changed its size.
*Added preferred hospital name and address fields to the patient edit
screen, and made some tweaks to better support going to and from the map
screen. In this case, I basically made it so that changes made to the
preferred hospital information in the map screen are reflected on return
to edit patients.

Things to do:
*Basic Map UI is extremely rough, but functional. Theres a lot of
polishing to be done, but I'll leave that for another Issue.
*I tried testing as much as possible, but I'm gonna need testing from
other eyes, I GUARANTEE you something will go wrong.